# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.CartItem:
    properties:
      product:
        $ref: '#/definitions/domain.Product'
      quantity:
        type: integer
    type: object
  domain.Product:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
    type: object
  dto.AddToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
      user_id:
        description: Optional, can be set from context in service layer
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  dto.RemoveFromCartRequest:
    properties:
      product_id:
        type: integer
      user_id:
        description: Optional, can be set from context in service layer
        type: integer
    required:
    - product_id
    type: object
...
  dto.UpdateCartItemRequest:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
      user_id:
        description: Optional, can be set from context in service layer
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  pkg.ErrorResponseSwag:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  pkg.SuccessResponseSwag:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API cho hệ thống e-commerce
  termsOfService: http://example.com/terms/
  title: E-commerce API
  version: "1.0"
paths:
...
  /cart/add:
    post:
      consumes:
      - application/json
      description: Add a product to the user's cart with specified quantity
      parameters:
      - description: Add to cart request
        in: body
        name: addToCartRequest
        required: true
        schema:
          $ref: '#/definitions/dto.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.SuccessResponseSwag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/clear:
    post:
      consumes:
      - application/json
      description: Remove all items from the user's cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.SuccessResponseSwag'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      summary: Clear cart
      tags:
      - Cart
  /cart/items:
    get:
      consumes:
      - application/json
      description: Retrieve the list of items in the user's cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pkg.SuccessResponseSwag'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.CartItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Get cart items
      tags:
      - Cart
  /cart/remove:
    post:
      consumes:
      - application/json
      description: Remove a product from the user's cart
      parameters:
      - description: Remove from cart request
        in: body
        name: removeFromCartRequest
        required: true
        schema:
          $ref: '#/definitions/dto.RemoveFromCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.SuccessResponseSwag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
  /cart/update:
    put:
      consumes:
      - application/json
      description: Update the quantity of a product in the user's cart
      parameters:
      - description: Update cart item request
        in: body
        name: updateCartItemRequest
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.SuccessResponseSwag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      summary: Update cart item quantity
      tags:
      - Cart
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"